test(fruit/domain): moved test files inside test folder
