+++ test/modules/fruit/domain/fruitId.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from '@jest/globals';
-import FruitId from './FruitId';
-import UniqueEntityID from '../../../shared/domain/UniqueEntityID';
+import FruitId from '../../../../src/modules/fruit/domain/FruitId';
+import UniqueEntityID from '../../../../src/shared/domain/UniqueEntityID';
 
 describe('FruitId ', () => {
   it('should be able to create FruitId successfully', async () => {
